feat: chat button now links to chat page feat: moved navbar to layout.tsx
