+++ apps/desktop/src/app/chat/page.tsx
@@ -1,10 +1,8 @@
-//import { Chat } from '../components/chat';
-import { Navbar } from '../components/navbar';
 
 export default function Home() {
 	return (
 		<div>
-			<Navbar />
+
 		</div>
 	); //Amaan
 }
