--- apps/desktop/src/app/components/chat/chat-button.tsx
@@ -1,7 +1,0 @@
-export const ChatButton = () => {
-	return (
-		<button className="bg-[#3399FF] text-[white] w-full py-4">
-			New Chat +
-		</button>
-	);
-}
