+++ apps/desktop/src/app/settings/page.tsx
@@ -1,10 +1,9 @@
-//import { Chat } from '../components/chat';
-import { Navbar } from '../components/navbar';
+
 
 export default function Home() {
 	return (
 		<div>
-			<Navbar />
+
 			<h1> "Hello world" </h1>
 		</div>
 	);
